move pages to containers folder and remove pages that are blank
