--- src/pages/Dragons.js
@@ -1,7 +1,0 @@
-import React from 'react';
-
-const Dragons = () => (
-  <div className="coming-soon">Dragons - Coming Soon...</div>
-);
-
-export default Dragons;
