--- src/pages/Landpads.js
@@ -1,7 +1,0 @@
-import React from 'react';
-
-const Landpads = () => (
-  <div className="coming-soon">Landpads- Coming Soon...</div>
-);
-
-export default Landpads;
