--- src/pages/Rockets.js
@@ -1,7 +1,0 @@
-import React from 'react';
-
-const Rockets = () => (
-  <div className="coming-soon">Rockets - Coming Soon...</div>
-);
-
-export default Rockets;
