--- src/pages/Ships.js
@@ -1,5 +1,0 @@
-import React from 'react';
-
-const Ships = () => <div className="coming-soon">Ships - Coming Soon...</div>;
-
-export default Ships;
